function that returns the sum of all its parameters

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -11,7 +11,7 @@ int sum_them_all(const unsigned int n, ...)
 {
 
 	va_list argm;
-	double sum = 0.0;
+	int sum = 0;
 	unsigned int i;
 
 	va_start(argm, n);
